Parsing changes

+++ TheWeatherTurtle/Models/JSON/WeatherDetails/Main.swift
@@ -8,7 +8,7 @@
 
 struct Main: Codable {
 	let temp: Double?
-	let pressure: Int?
+	let pressure: Double?
 	let humidity: Int?
 	let temp_min: Double?
 	let temp_max: Double?
