+++ TheWeatherTurtle/Models/JSON/WeatherDetails/Wind.swift
@@ -8,5 +8,5 @@
 
 struct Wind: Codable {
 	let speed: Double?
-    let deg: Int?
+    let deg: Double?
 }
